feat(mongodb-mongoose) Export logic from controller into a model to interact with DB
