+++ router/tourRouter.js
@@ -3,7 +3,7 @@ const tourController = require('../controllers/tourController');
 
 const router = express.Router();
 
-router.param('id', tourController.checkID)
+// router.param('id', tourController.checkID)
 
 // Tour Routes
 router.route('/').get(tourController.getAllTours).post(tourController.checkBody, tourController.createTour);
