+++ server.js
@@ -13,36 +13,18 @@ mongoose.connect(DB, {
   useUnifiedTopology: true
 }).then(con => console.log('DB connection successful!'))
 
-const tourSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: true,
-    unique: true
-  },
-  rating: {
-    type: Number,
-    default: 4.5
-  },
-  price: {
-    type: Number,
-    required: [true, 'A tour must have a price']
-  }
-})
+// const newTour = new Tour({
+//   name: 'The Mushroom Adventure',
+//   price: 497,
+// })
 
-const Tour = mongoose.model('Tour', tourSchema);
-
-const newTour = new Tour({
-  name: 'The Mushroom Adventure',
-  price: 497,
-})
-
-newTour.save()
-  .then(doc => {
-    console.log(doc);
-  })
-  .catch(err => {
-    console.log('Error saving tour: ', err)
-  })
+// newTour.save()
+  // .then(doc => {
+  //   console.log(doc);
+  // })
+  // .catch(err => {
+  //   console.log('Error saving tour: ', err)
+  // })
 
 
 const PORT = process.env.PORT || 3000;
